+++ resources/views/elements/question.blade.php
@@ -7,6 +7,9 @@
         @if(!empty($options))
             <span>[<span class="text-emerald-500">{{implode('/',$options)}}</span>]</span>
         @endif
+        @if(!is_null($default ?? null))
+            <span class="text-stone-400">(default: {{$default}})</span>
+        @endif
     </div>
     <div>❯</div>
 </div>