+++ tests/TestCase.php
@@ -2,6 +2,7 @@
 
 namespace OmniTerm\Tests;
 
+use OmniTerm\OmniTermServiceProvider;
 use Orchestra\Testbench\TestCase as Orchestra;
 
 class TestCase extends Orchestra
@@ -9,6 +10,12 @@ class TestCase extends Orchestra
     protected function setUp(): void
     {
         parent::setUp();
+    }
 
+    protected function getPackageProviders($app): array
+    {
+        return [
+            OmniTermServiceProvider::class,
+        ];
     }
 }